Refactor/dev

See merge request react-node-track/class01/web-project1/team04/help_me_backend!20

+++ src/controller/user.controller.js
@@ -47,6 +47,8 @@ export const userController = {
       const userId = req.user.id;
       const user = await userService.getFindUser(userId);
 
+      if (!userId) throw new Error ('Bad Request+유저 정보를 찾을 수 없음');
+
       res.status(200).json(user);
     } catch(e) {
       next(e);
@@ -101,6 +103,9 @@ export const userController = {
   deleteUser: async (req, res, next) => {
     try {
       const email = req.user.email;
+
+      if (!email) throw new Error ('Bad Request+유저 정보를 찾을 수 없음');
+
       const user = await userService.deleteUser(email);
       res.status(200).json(user);
     } catch(e) {
